update standard e2e tests to run isolated
