+++ package.json
@@ -56,7 +56,7 @@
 		"babel-loader": "^9.1.2",
 		"core-js": "^3.30.0",
 		"css-loader": "^6.7.3",
-		"cypress": "8.5.0",
+		"cypress": "^12.7.0",
 		"eslint": "^8.37.0",
 		"eslint-plugin-react": "^7.32.2",
 		"file-loader": "^6.2.0",
