+++ tests/cypress/support/commands.js
@@ -40,7 +40,15 @@ Cypress.Commands.add('addCloudSnap', (branch = 'production') => {
 	cy.addScript(`https://snapui.searchspring.io/${packageJSON.searchspring.siteId}/${branch}/bundle.js`);
 });
 
-Cypress.Commands.add('snapController', (controllerId = 'search') => {
+const defaultOptions = {
+	delay: 200,
+}
+
+Cypress.Commands.add('snapController', (controllerId = 'search', options) => {
+
+	const mergedOptions = {...defaultOptions, ...options};
+	cy.wait(mergedOptions.delay); 
+	
 	return cy.window().then((window) => {
 		return new Cypress.Promise((resolve) => {
 			const checkTimeout = 100;
